FCT2-20332 extract BaseTransferMaterialsTab to cut duplication
SonarQube flagged duplicated lines across the two page objects. Move the
Egress-side and common helpers that are identical on both screens
(selectAllEgressFiles, selectEgressFileByName, navigateToFolder,
waitForEgressFileByName, sortNetAppByDateDescending, the fixture-missing
error) into an abstract BaseTransferMaterialsTab that both extend, and
dedupe the repeated error-route regex and file-checkbox locator inside
TransferMaterialsTabV1. Behaviour unchanged; tsc clean.
