+++ e2e/pw/pages/TransferMaterialsTabV1.ts
@@ -1,6 +1,12 @@
-import { Page, expect } from "@playwright/test";
+import { expect, Locator } from "@playwright/test";
 import { TransferMaterialsTabApi } from "./TransferMaterialsTabApi";
 import { TransferDestinationPage } from "./TransferDestinationPage";
+import { BaseTransferMaterialsTab } from "./BaseTransferMaterialsTab";
+
+// The new screen navigates to one of these routes on a failed or
+// duplicate-rejected transfer (used to detect + recover from the error page).
+const TRANSFER_ERROR_ROUTE =
+  /\/case\/\d+\/case-management\/(transfer-errors|transfer-permissions-error|transfer-resolve-file-path)/;
 
 /**
  * New-screen (v1) Transfer Materials page object, selected by
@@ -9,17 +15,22 @@ import { TransferDestinationPage } from "./TransferDestinationPage";
  * `Copy selected` / `Move selected` buttons; direction toggles via a
  * `View Shared Drive` / `View Egress` link; no confirm modal (Copy/Move navigate
  * to a destination-tree page, driven by `TransferDestinationPage`); errors use
- * the new `transfer-errors` / `transfer-permissions-error` /
- * `transfer-resolve-file-path` routes. The Egress table is unchanged.
+ * the routes above. Egress-side helpers come from `BaseTransferMaterialsTab`.
  */
-export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
-  private readonly page: Page;
+export class TransferMaterialsTabV1
+  extends BaseTransferMaterialsTab
+  implements TransferMaterialsTabApi
+{
+  protected readonly netAppWrapperTestId = "shared drive-table-wrapper";
   // Remembered so confirmTransfer knows whether the destination-page button
   // reads "Copy to …" or "Move to …".
   private lastAction: "Copy" | "Move" = "Copy";
 
-  constructor(page: Page) {
-    this.page = page;
+  /** Row checkbox for a shared-drive file, by its exact aria-label. */
+  private fileCheckbox(fileName: string): Locator {
+    return this.page
+      .locator(`input[type="checkbox"][aria-label="select file ${fileName}"]`)
+      .first();
   }
 
   /** Click the shared Copy/Move control. Renders in a top and bottom bar (target
@@ -60,13 +71,6 @@ export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
       .click();
   }
 
-  async selectAllEgressFiles(): Promise<void> {
-    await this.page
-      .getByTestId("egress-table-wrapper")
-      .getByLabel("Select folders and files")
-      .check();
-  }
-
   async selectNetAppFiles(indices: number[]): Promise<void> {
     // Select per-row checkboxes by aria-label (page-scoped, forced). Only the
     // shared-drive panel shows while it's the source, so these are its rows;
@@ -82,27 +86,6 @@ export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
   }
 
   /**
-   * Sort the shared-drive panel by last-modified date descending (two header
-   * clicks). Best-effort: the date sort is unreliable and source identity
-   * doesn't matter, so a flaky/slow header is tolerated — proceed unsorted.
-   */
-  async sortNetAppByDateDescending(): Promise<void> {
-    const header = this.page
-      .getByTestId("shared drive-table-wrapper")
-      .getByRole("button", { name: "Last modified date" });
-    try {
-      await header.click({ timeout: 20_000 });
-      await this.waitForNetAppFiles();
-      await header.click({ timeout: 20_000 });
-      await this.waitForNetAppFiles();
-    } catch {
-      console.log(
-        "  [sortNetAppByDateDescending] date sort skipped (best-effort)",
-      );
-    }
-  }
-
-  /**
    * Select a shared-drive source row by exact filename; throws a clear
    * fixture-missing message if absent (see `tests/seed-netapp-fixture.setup.ts`).
    */
@@ -111,31 +94,14 @@ export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
     // shared-drive rows can be in the DOM but not "visible" to Playwright,
     // source switches don't reliably re-trigger the loader, and the listing
     // can take well over 30s — so a visibility gate false-negatives.
-    const checkbox = this.page
-      .locator(`input[type="checkbox"][aria-label="select file ${fileName}"]`)
-      .first();
+    const checkbox = this.fileCheckbox(fileName);
     try {
       await checkbox.waitFor({ state: "attached", timeout: 120_000 });
     } catch {
-      throw new Error(
-        `Required NetApp fixture '${fileName}' not found in the shared drive panel.\n` +
-          `  Seed it via:\n` +
-          `    bash:       RUN_SEED=1 npx playwright test --project=seed-netapp-fixture\n` +
-          `    powershell: $env:RUN_SEED=1; npx playwright test --project=seed-netapp-fixture\n` +
-          `  See README "Required NetApp fixture" for details.`,
-      );
+      throw this.fixtureMissingError(fileName, "shared drive panel");
     }
     await checkbox.scrollIntoViewIfNeeded().catch(() => {});
     await checkbox.check({ force: true });
-  }
-
-  async selectEgressFileByName(fileName: string): Promise<void> {
-    const row = this.page
-      .getByTestId("egress-table-wrapper")
-      .locator("tbody tr", { hasText: fileName });
-    const checkbox = row.locator('input[type="checkbox"]');
-    await checkbox.scrollIntoViewIfNeeded();
-    await checkbox.check();
   }
 
   async selectAction(action: "Copy" | "Move"): Promise<void> {
@@ -166,15 +132,13 @@ export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
     const successBanner = this.page.getByTestId(
       "transfer-success-notification-banner",
     );
-    const errorRoute =
-      /\/case\/\d+\/case-management\/(transfer-errors|transfer-permissions-error|transfer-resolve-file-path)/;
 
     await Promise.race([
       successBanner.waitFor({ state: "visible", timeout }),
-      this.page.waitForURL(errorRoute, { timeout }),
+      this.page.waitForURL(TRANSFER_ERROR_ROUTE, { timeout }),
     ]);
 
-    if (errorRoute.test(this.page.url())) {
+    if (TRANSFER_ERROR_ROUTE.test(this.page.url())) {
       const detail = await this.page
         .locator("main")
         .innerText()
@@ -193,9 +157,7 @@ export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
    * be re-entered. No-op otherwise.
    */
   async dismissTransferErrorIfPresent(): Promise<void> {
-    const errorRoute =
-      /\/case\/\d+\/case-management\/(transfer-errors|transfer-permissions-error|transfer-resolve-file-path)/;
-    if (!errorRoute.test(this.page.url())) return;
+    if (!TRANSFER_ERROR_ROUTE.test(this.page.url())) return;
     await this.page.getByRole("link", { name: "Back", exact: true }).click();
   }
 
@@ -210,49 +172,6 @@ export class TransferMaterialsTabV1 implements TransferMaterialsTabApi {
   ): Promise<void> {
     await this.switchToNetAppSource();
     await this.waitForNetAppFiles();
-    const checkbox = this.page
-      .locator(`input[type="checkbox"][aria-label="select file ${fileName}"]`)
-      .first();
-    await checkbox.waitFor({ state: "attached", timeout });
-  }
-
-  async navigateToFolder(folderName: string): Promise<void> {
-    await this.page.getByRole("button", { name: folderName }).click();
-  }
-
-  /**
-   * Wait for a file to appear in the Egress panel. Egress doesn't auto-refresh,
-   * so each retry does a full `page.reload()` (busts the React Query cache;
-   * in-app re-navigation serves the stale listing) and re-navigates the folder
-   * path. The transfer route + Egress-as-source survive the reload.
-   */
-  async waitForEgressFileByName(
-    fileName: string,
-    folderPath: string[],
-    timeout: number = 300_000,
-  ): Promise<void> {
-    const egressTable = this.page.getByTestId("egress-table-wrapper");
-    const start = Date.now();
-
-    const fileVisible = async () =>
-      (await egressTable.locator("tbody tr", { hasText: fileName }).count()) >
-      0;
-
-    if (await fileVisible()) return;
-
-    while (Date.now() - start < timeout) {
-      console.log(`  Waiting for ${fileName} to be indexed; refreshing...`);
-      await this.page.waitForTimeout(5_000);
-      await this.page.reload();
-      await this.waitForEgressFiles();
-      for (const folder of folderPath) {
-        await this.navigateToFolder(folder);
-        await this.waitForEgressFiles();
-      }
-      if (await fileVisible()) return;
-    }
-    throw new Error(
-      `Timed out waiting for ${fileName} to appear in Egress panel (timeout: ${timeout}ms)`,
-    );
+    await this.fileCheckbox(fileName).waitFor({ state: "attached", timeout });
   }
 }
